Rename shared to clients and utils
