+++ src/datarobot_genai/drmcp/core/dynamic_tools/deployment/register.py
@@ -16,10 +16,10 @@ import logging
 import datarobot as dr
 from fastmcp.tools.tool import Tool
 
+from datarobot_genai.drmcp.core.clients import get_api_client
 from datarobot_genai.drmcp.core.dynamic_tools.deployment.config import create_deployment_tool_config
 from datarobot_genai.drmcp.core.dynamic_tools.register import register_external_tool
 from datarobot_genai.drmcp.core.exceptions import DynamicToolRegistrationError
-from datarobot_genai.drmcp.core.shared import get_api_client
 
 logger = logging.getLogger(__name__)
