+++ src/datarobot_genai/drmcp/tools/predictive/training.py
@@ -21,8 +21,8 @@ from dataclasses import dataclass
 
 import pandas as pd
 
+from datarobot_genai.drmcp.core.clients import get_sdk_client
 from datarobot_genai.drmcp.core.mcp_instance import dr_mcp_tool
-from datarobot_genai.drmcp.core.shared import get_sdk_client
 
 logger = logging.getLogger(__name__)
